bd/requirements.txt
sqlalchemy
asyncio
asyncpg
python-dotenv
fastapi
uvicorn
pydantic
cryptography
httpx

